+++ jest.setup.js
@@ -1,8 +1,8 @@
+const { createCanvas } = require('@napi-rs/canvas');
 const { JSDOM } = require('jsdom');
 // new TextDecoder('ascii') is not supported in node12
 // use TextDecoder and TextEncoder from `text-encoding`
 const { TextDecoder, TextEncoder } = require('text-encoding');
-const { createCanvas } = require('@napi-rs/canvas');
 
 const jsdom = new JSDOM('<div id="main"></div>', {
   // https://github.com/jsdom/jsdom#basic-options
