+++ packages/terminal-next/src/browser/terminal.hover.manager.ts
@@ -1,9 +1,7 @@
 import { Injectable } from '@opensumi/di';
 import { Disposable } from '@opensumi/ide-core-common';
 
-import { ITerminalHoverManagerService } from '../common';
-
-import { ILinkHoverTargetOptions } from './links/link-manager';
+import { ITerminalHoverManagerService, ILinkHoverTargetOptions } from '../common';
 
 const TIPS_OFFSET_Y = 20;
 const TIPS_OFFSET_X = 5;
